frontend/requirements.txt
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dash==1.20.0
dash-bootstrap-components==0.12.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
plotly==4.14.3
polygon-api-client==0.1.9
python-dotenv==0.17.0
requests==2.25.1
retrying==1.3.3
six==1.15.0
urllib3==1.26.4
waitress==2.0.0
websocket-client==0.58.0
websockets==8.1
Werkzeug==1.0.1
